user dashboard get and put APi updated
