+++ codebase/models/api/user.js
@@ -53,6 +53,7 @@ const UsersSchema = new mongoose.Schema({
   username: {
     type: String,
     required: true,
+    unique: true,
     trim: true,
     minlength: 3,
     maxlength: 20,
